requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
email-validator==2.1.0
Flask-Bcrypt==1.0.1
SQLAlchemy==1.4.46
Werkzeug==2.3.7
WTForms==3.1.0
Pillow==10.0.0
mysqlclient==2.2.0 